+++ packages/dialtone-vue3/recipes/conversation_view/message_input/message_input.vue
@@ -162,6 +162,7 @@
                 @add-emoji="$emit('add-emoji')"
                 @skin-tone="onSkinTone"
                 @selected-emoji="(emoji) => onSelectEmoji(emoji, close)"
+                @scroll-bottom-reached="$emit('emoji-scroll-bottom-reached')"
               />
             </template>
           </dt-popover>
@@ -745,6 +746,12 @@ export default {
      * @type {Boolean}
      */
     'add-emoji',
+
+    /**
+     * Emitted when the emoji picker scroll reaches the bottom
+     * @event emoji-scroll-bottom-reached
+     */
+    'emoji-scroll-bottom-reached',
   ],
 
   data () {