+++ packages/dialtone-vue3/recipes/conversation_view/message_input/message_input_default.story.vue
@@ -53,6 +53,7 @@
       @paste-media="$attrs.onPasteMedia"
       @notice-close="$attrs.onNoticeClose"
       @cancel="$attrs.onCancel"
+      @emoji-scroll-bottom-reached="$attrs.onEmojiScrollBottomReached"
     >
       <template
         v-if="$attrs.emojiGiphyPicker"